lesson-13/application/requirements.txt
boto3==1.34.144
botocore==1.34.144
psycopg2-binary==2.9.9
pytest==7.4.4
pytest-mock==3.12.0
moto==4.2.14
